fix: change isize arg to usize, string to char

+++ src/compiler/utils/error.rs
@@ -64,7 +64,7 @@ impl CompileError {
             column: start.column,
         });
         let frame = start
-            .map(|start| get_code_frame(props.source.to_string(), (start.line - 1) as isize, start.column as isize));
+            .map(|start| get_code_frame(props.source.to_string(), start.line - 1, start.column));
 
         Self {
             code: props.code.to_string(),
